single lambda dispatch and campaignCache as a Task
