+++ src/workers/jobs.js
@@ -1081,35 +1081,6 @@ export async function loadMessages(csvFile) {
   });
 }
 
-export async function loadCampaignCache(
-  campaign,
-  organization,
-  { remainingMilliseconds }
-) {
-  // Asynchronously start running a refresh of all the campaign data into
-  // our cache.  This should refresh/clear any corruption
-  console.log("loadCampaignCache async tasks...", campaign.id);
-  const loadContacts = cacheableData.campaignContact
-    .loadMany(campaign, organization, { remainingMilliseconds })
-    .then(() => {
-      console.log("FINISHED contact loadMany", campaign.id);
-    })
-    .catch(err => {
-      console.error("ERROR contact loadMany", campaign.id, err, campaign);
-    });
-  const loadOptOuts = cacheableData.optOut.loadMany(organization.id);
-  const loadAssignments = cacheableData.campaignContact.updateCampaignAssignmentCache(
-    campaign.id
-  );
-
-  if (global.TEST_ENVIRONMENT) {
-    // otherwise this races with texting
-    await loadContacts;
-    await loadOptOuts;
-    await loadAssignments;
-  }
-}
-
 // Temporary fix for orgless users
 // See https://github.com/MoveOnOrg/Spoke/issues/934
 // and job-processes.js
